-<engine> 손바닥특징추출엔진을 hfir_1.2로 교체함. 손바닥이 포화되는 경우의 카메라조종론리를 수정함.

+++ solutions/usb_cam/application/FaceEngine/_dnn/dic_manage.cpp
@@ -636,7 +636,7 @@ int getDicChecSumChecked(int nMachineIndex)
 
     if(nMachineIndex == MachineFlagIndex_DNN_Feature_Hand)
     {
-        if(g_nDicCheckSum_FEAT_Hand == DNN_FEAT_HAND_CHECKSUM)
+        if((unsigned int)g_nDicCheckSum_FEAT_Hand == DNN_FEAT_HAND_CHECKSUM)
         {
             return 1;
         }
